Manage outliers from JDemetra+ workspaces

These functions allow extracting, exporting and importing outliers detected in a JDemetra+ workspace:

retrieve_outliers() extracts outliers from a .xml workspace.

export_outliers() saves extracted outliers into a YAML file.

import_outliers() loads outliers back to a .xml workspace from a YAML file.

They are useful for archiving and reusing outliers detected during seasonal adjustment workflows.

Usage

export_outliers(x, ws_name, path = NULL, verbose = TRUE)

import_outliers(ws_name, path = NULL, verbose = TRUE)

retrieve_outliers(
  jws,
  domain = TRUE,
  estimation = FALSE,
  point = FALSE,
  verbose = TRUE
)

Arguments

x

list A list of outliers, as returned by retrieve_outliers().

ws_name

[character] The name of the workspace, used to build default YAML filenames.

path

[character] Path to a YAML file to write to or read from. If missing, defaults to "regression/outliers_<ws_name>.yaml".

verbose

[logical] Whether to print informative messages (default: TRUE).

jws

A Java Workspace object, as returned by rjd3workspace::jws_open() or rjd3workspace::jws_new().

domain

Boolean indicating if the outliers should be extracted from the domain specification.

estimation

Boolean indicating if the outliers should be extracted from the estimation specification.

point

Boolean indicating if the outliers should be extracted from the point specification.

Value

retrieve_outliers() returns a named list where each element corresponds to a series in the workspace and contains the names of existing (detected or pre-specified) outliers.

export_outliers() invisibly returns the path of the YAML file written.

import_outliers() returns a list of outliers read from a YAML file.

Examples

library("rjd3workspace")
# Load a Workspace
file <- system.file("workspaces", "workspace_test.xml", package = "rjd3workspace")
jws <- jws_open(file)
#> Error in .jcall(obj = "jdplus/sa/base/workspace/Ws", returnSig = "Ljdplus/sa/base/workspace/Ws;",     method = "open", full_file_name): RcallMethod: cannot determine object class

my_dir <- tempdir()

outliers <- retrieve_outliers(jws, point = TRUE, domain = FALSE, estimation = FALSE)
#> Error in .jcall(jws, "V", "computeAll"): java.lang.NoClassDefFoundError: Could not initialize class jdplus.sa.base.workspace.Ws
export_outliers(outliers, ws_name = "workspace1", path = my_dir)
#> The outliers will be written at  /tmp/RtmpTFgV5B/outliers_workspace1.yaml 
#> Error: object 'outliers' not found
imported <- import_outliers(ws_name = "workspace1", path = my_dir)
#> Error in normalizePath(file.path(path, file_name), mustWork = TRUE): path[1]="/tmp/RtmpTFgV5B/outliers_workspace1.yaml": No such file or directory